Split tests so they can be run in parallel
