+++ test/index.spec/index.spec.ts
@@ -1,0 +1,4 @@
+import { it } from '@jest/globals';
+
+it.todo('thoroughly test cached pings');
+// TODO: test connecting to a Procedure's bound endpoint with a socket and throwing random data at it
